move embed metadata ops to runtime
