+++ src/runtime/vortex/runtime/onnx/graph_ops/embed_class_names_metadata.py
