+++ src/runtime/vortex/runtime/onnx/graph_ops/embed_metadata.py
