+++ src/runtime/vortex/runtime/onnx/graph_ops/embed_model_property.py
