+++ src/runtime/vortex/runtime/onnx/graph_ops/embed_output_format_metadata.py
