+++ src/runtime/vortex/runtime/onnx/graph_ops/helper.py
@@ -1,0 +1,7 @@
+import onnx
+
+def get_metadata_prop(model: onnx.ModelProto, key: str) -> onnx.onnx_pb.StringStringEntryProto:
+    for prop in model.metadata_props:
+        if prop.key == key:
+            return prop
+    return None
